Move ParkingSpaceController to Frontend subfolder
